# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ngx-admin Backend Bundle API",
    "description": "ngx-admin Backend Bundle API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Bearer <my own JWT token>"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User Authentication returning a JWT",
        "description": "Authenticate a user",
        "parameters": [
          {
            "name": "loginDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/loginDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          }
        }
      }
    },
    "/auth/sign-up": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign up new user",
        "description": "Sign up new user",
        "parameters": [
          {
            "name": "signUpDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/signUpDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          }
        }
      }
    },
    "/auth/reset-pass": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Change password",
        "description": "Change password for current user",

        "parameters": [
          {
            "name": "changePasswordDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/changePasswordDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error message",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/request-pass": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request password",
        "description": "Request password by email",
        "parameters": [
          {
            "name": "requestPasswordDto",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error message",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/refresh-token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh user token",
        "description": "Refresh user token",
        "parameters": [
          {
            "name": "refreshTokenDTO",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error message",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "loginDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "signUpDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "lang": {
          "type": "string"
        },
        "create_by": {
          "type": "string"
        }
      }
    },
    "changePasswordDto": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      }
    },
    "Token": {
      "type": "object",
      "properties": {
        "token": {
          "type": "object",
          "properties": {
            "expires_in": {
              "type": "integer"
            },
            "access_token": {
              "type": "string"
            },
            "refresh_token": {
              "type": "string"
            }
          }
        }
      }
    },
...
  }
}